[TASK-211] <<fixed>> rename BadgeBanner -> BadgePageBanner
