+++ itp-front/src/components/BadgePageBanner.vue
